deploy_package/backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
django-cors-headers
Pillow
gunicorn
whitenoise
